Extract envelope queue classes to a separate package
